docker/copy_docker/requirements.txt
flask
torch
torchvision
Pillow
prometheus_client
